requirements.txt
pandas
numpy
nltk
wordcloud
scikit-learn
pyyaml
